# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthInfo": {
        "additionalProperties": false,
        "properties": {
          "local": {
            "$ref": "#/components/schemas/LocalAuthInfo"
          }
        },
        "type": "object"
      },
      "AuthTokenBodyBody": {
        "additionalProperties": false,
        "properties": {
          "token": {
            "description": "The short-lived JWT auth token. Send it as a bearer token on subsequent requests.",
            "readOnly": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
...
      "LegalInfo": {
        "additionalProperties": false,
        "properties": {
          "imprint_url": {
            "type": "string"
          },
          "privacy_policy_url": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "LocalAuthInfo": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "registration_enabled": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "Login": {
        "additionalProperties": false,
        "properties": {
          "long_token": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "TaskBoardErrorModel": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "taskboard numeric error code; see https://taskboard.io/docs/errors/",
            "format": "int64",
            "readOnly": true,
            "type": "integer"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
      "TaskBoardInfos": {
        "additionalProperties": false,
        "properties": {
          "allow_icon_changes": {
            "description": "Whether users may change project icons.",
            "type": "boolean"
          },
          "auth": {
            "$ref": "#/components/schemas/AuthInfo",
            "description": "The authentication methods enabled on this instance."
          },
          "concurrent_writes": {
            "description": "Whether the configured database supports concurrent writes. False on SQLite; clients should serialize batched writes when this is false.",
            "type": "boolean"
          },
          "demo_mode_enabled": {
            "description": "Whether this instance runs in demo mode (data is periodically reset).",
            "type": "boolean"
          },
          "email_reminders_enabled": {
            "description": "Whether email reminders are enabled.",
            "type": "boolean"
          },
          "frontend_url": {
            "description": "The publicly configured frontend URL of this instance.",
            "type": "string"
          },
          "legal": {
            "$ref": "#/components/schemas/LegalInfo",
            "description": "Links to the instance's legal documents."
          },
          "max_file_size": {
            "description": "The maximum allowed upload size, as a human-readable string (e.g. 20MB).",
            "type": "string"
          },
          "max_items_per_page": {
            "description": "The maximum number of items a paginated endpoint returns per page.",
            "format": "int64",
            "type": "integer"
          },
          "motd": {
            "description": "The message of the day, shown to all users.",
            "type": "string"
          },
          "public_teams_enabled": {
            "description": "Whether public teams are enabled.",
            "type": "boolean"
          },
          "task_attachments_enabled": {
            "description": "Whether task attachments are enabled.",
            "type": "boolean"
          },
          "task_comments_enabled": {
            "description": "Whether task comments are enabled.",
            "type": "boolean"
          },
          "user_deletion_enabled": {
            "description": "Whether users may delete their own account.",
            "type": "boolean"
          },
          "version": {
            "description": "The taskboard version this instance runs.",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "JWTKeyAuth": {
        "bearerFormat": "JWT",
        "description": "User session JWT issued via /api/v1/login.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "taskboard API",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/info": {
      "get": {
        "description": "Returns version, frontend URL, motd and the enabled features of this taskboard instance. Public — no authentication required.",
        "operationId": "info",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskBoardInfos"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskBoardErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [],
        "summary": "Instance info",
        "tags": [
          "service"
        ]
      }
    },
...
    "/login": {
      "post": {
        "description": "Logs a user in with username and password (and a TOTP passcode when 2FA is enabled), returning a short-lived JWT. A long-lived refresh token is set as an HttpOnly cookie scoped to the refresh endpoint.",
        "operationId": "auth-login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenBodyBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskBoardErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [],
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/user/token/refresh": {
      "post": {
        "description": "Exchanges the refresh-token cookie for a new short-lived JWT. The refresh token is rotated on every call, so the previous one stops working. A new HttpOnly refresh cookie is set on the response.",
        "operationId": "auth-refresh-token",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenBodyBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskBoardErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [],
        "summary": "Refresh user token",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "security": [
    {
      "JWTKeyAuth": []
    }
  ],
  "servers": [
    {
      "url": "/api/v1"
    }
  ]
}